refactor: change UniInfoEditor to MarkdownEditor, moved to /ui & clean forms
