+++ client/components/uni-info/index.ts
@@ -4,7 +4,6 @@ import UniInfoForm from './uni-info-form'
 import UniInfoLoading from './uni-info-loading'
 import UniInfoMobileMenu from './uni-info-mobile-menu'
 import UniInfoNav from './uni-info-nav'
-import UniInfoEditor from './uni-info-editor'
 
 export {
   UniInfoContent,
@@ -13,5 +12,4 @@ export {
   UniInfoLoading,
   UniInfoMobileMenu,
   UniInfoNav,
-  UniInfoEditor,
 }
